Add data_storage.py, for access to datafiles without authentication
This commit adds data_storage.py, which implements just
enough of the Google Cloud Storage Python API to access the
Deepmind/Graphcast data bucket through the HTTPS interface,
without authentication.  This should make it easier to run
the graphcast_demo Jupyter notebook on non-Colab
environments.

The Jupyter notebook now checks for import errors when
importing google.colab and google.cloud; an ImportError from
either will trigger the use of this module.

The module downloads data as necessary.  If tqdm is present
in the environment, it will use that library to generate a
progress bar.

If run as an independent script, data_storage.py will
download all files of the `dm_graphcast` bucket (or another
public bucket specified on the command line), making it
possible to run the Jupyter notebook in a compute
environment where a live Internet connection is either
unavailable or undesired (e.g. a job queue).